use ngspice specific utility functions

+++ src/maths/KLU/klu.c
@@ -116,11 +116,11 @@ size_t KLU_kernel_factor            /* 0 if failure, size of LU if OK */
     {
         Lsize = -Lsize ;
         Lsize = MAX (Lsize, 1.0) ;
-        lsize = (int) Lsize * anz + n ;
+        lsize = Lsize * anz + n ;
     }
     else
     {
-        lsize = (int) Lsize ;
+        lsize = Lsize ;
     }
 
     usize = lsize ;
@@ -130,8 +130,8 @@ size_t KLU_kernel_factor            /* 0 if failure, size of LU if OK */
 
     maxlnz = (((double) n) * ((double) n) + ((double) n)) / 2. ;
     maxlnz = MIN (maxlnz, ((double) Int_MAX)) ;
-    lsize  = MIN ((int) maxlnz, lsize) ;
-    usize  = MIN ((int) maxlnz, usize) ;
+    lsize  = MIN (maxlnz, lsize) ;
+    usize  = MIN (maxlnz, usize) ;
 
     PRINTF (("Welcome to klu: n %d anz %d k1 %d lsize %d usize %d maxlnz %g\n",
         n, anz, k1, lsize, usize, maxlnz)) ;
